Heroku changes

+++ client/src/pages/Main.js
@@ -31,10 +31,10 @@ class Main extends Component {
     }
 
       handleDeleteButton = (event) => {
-        API.deleteFood({ id: event.target.id }).then(this.getPantryItems());
+        API.deleteFood({ id: event.target.id }).then(this.getPantryItems(this.state.user_id));
       }
 
-      renderItems = () => (this.state.pantryItems.length > 0 ?
+      renderItems = () => (this.state.pantryItems.length > 0 && this.state.user_id ?
         this.state.pantryItems.map(item => (
           <PantryCard
             id={item.id}
@@ -42,7 +42,7 @@ class Main extends Component {
             date={item.purchase_date}
             handleDeleteButton={this.handleDeleteButton}
           />
-        )) : <h5>Save some articles here!</h5>
+        )) : <h5>Upload a receipt to get started!</h5>
       )
 
       render() {
